+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add playbook for generating deploy keys for environments
+
 ### Changed
 - Upgrade Next.js to 13.1 (@marteinn)